requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.2
blinker==1.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
ecdsa==0.18.0
fastapi==0.86.0
filelock==3.12.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.3
more-itertools==8.10.0
oauthlib==3.2.0
passlib==1.7.4
platformdirs==3.5.1
pyasn1==0.5.0
pydantic==1.10.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==5.4.1
requests==2.31.0
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.20.4
typing_extensions==4.4.0
unattended-upgrades==0.1
urllib3==2.0.4
uvicorn==0.19.0
virtualenv==20.23.0
wadllib==1.3.6
webencodings==0.5.1
zipp==1.0.0
